Add functionality to restart app after turning on FSM
This update revolves around the TurnFsmCommand in the clio Command namespace. Two new variables were added: IApplicationClient and EnvironmentSettings. The primary change is the addition of a check if .NetCore is turned on, and if so, the app is restarted after the file system mode (FSM) has been turned on.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>