Fsm fix (#261)
* Add functionality to restart app after turning on FSM

This update revolves around the TurnFsmCommand in the clio Command namespace. Two new variables were added: IApplicationClient and EnvironmentSettings. The primary change is the addition of a check if .NetCore is turned on, and if so, the app is restarted after the file system mode (FSM) has been turned on.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>

* Add command name to success log message

Two changes were made to enhance logging capabilities in the RemoteCommand class. Firstly, an import modification was made, introducing three new assemblies. Secondly, the Execute method now includes the command name in the logging information. This enhances readability and provides additional context for the log messages.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>

---------

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>