Add RemoveCreatio command and implement IIS site scanning
This commit adds the 'RemoveCreatio' command and implements the scanning of IIS sites. The scanning component uses the IAppCmd interface to interact with appcmd.exe and collect site binding and type information. This information is used to identify and classify Creatio-related sites by their bindings, physical paths and database types.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>