Add logger to SysSettingsManager and refactor SysSettings functionality
The SysSettingsManager now takes an ILogger in the constructor to log operations. A method has been added to create a SysSetting if it doesn't exist, improving efficiency. Code refactoring includes making function names more descriptive. This includes changing the function "GetSysSettingType" to "GetSysSettingByCode" which provides a clearer indication of its purpose.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>