Implement save-state command for WebService state
The commit introduces a "save-state" command that saves the state of a Creatio instance, specifically web services, to a file. It also includes test cases for this functionality. The implementation includes creating or updating the necessary classes and data structures to retrieve web service states and serialize them to a YAML file.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>