Update feature handling in SaveSettingsToManifestCommand
Refactored how features are fetched and handled in the SaveSettingsToManifestCommand. Previously commented-out code to get feature values has been enabled, and the mechanism for fetching and handling the feature state has been updated to accommodate multiple states for a single feature. In case of duplicate states, entries are distinguished by appending a unique identifier.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>