Update package dependencies to latest versions (#288)
The commit updates the version of several package dependencies in the solution. From ATF.Repository and Autofac to YamlDotNet, all have been bumped up to their latest versions in both clio.csproj and clio.tests.csproj files. This is part of an ongoing maintenance effort to keep the software up-to-date and take advantage of recent bug fixes, improvements, and features offered by these dependencies.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>