Update handling of working directories and disable environment ping command
This commit includes changes on how working directories are executed upon checking if its value is null. If not null, the `InternalExecute` function is now directly called. In addition, the execution of the ping command and the application of environment manifest options in the cloning environment process are temporarily disabled for debugging purposes.

Signed-off-by: Kirill Krylov <k.krylov@creatio.com>